requirements.txt
streamlit==1.28.1
volcengine-python-sdk[ark]
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.1
plotly==5.17.0
pandas==2.1.3
numpy==1.24.3 